TAMAYA-382 link updates
This PR makes 2 changes, per the February 2019 report.

1.  Updates artifact download links to use the mirror lookup link or
archive link as makes sense.
2.  Updates mail-archive link to work around problems putting the '@'
symbol in links in Asciidoc, but with the Apache list server not
accepting any encoded values in the URL.

I also changed a reference from apache gitbox over to github for new
devs, which seems in line with the move we made last year.